requirements-server.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-server.in
#
aioredis==1.3.1
    # via channels-redis
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==4.0.2
    # via
    #   aioredis
    #   redis
attrs==22.2.0
    # via
    #   automat
    #   jsonschema
    #   service-identity
    #   twisted
autobahn==23.1.2
    # via daphne
automat==22.10.0
    # via twisted
azure-core==1.26.3
    # via azure-storage-blob
azure-storage-blob==12.15.0
    # via django-storages
billiard==3.6.4.0
    # via celery
boto3==1.26.79
    # via -r requirements-server.in
botocore==1.29.79
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via
    #   -r requirements-server.in
    #   django-celery-results
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
chainmap==1.0.3
    # via -r requirements-server.in
channels==2.4.0
    # via
    #   -r requirements-server.in
    #   channels-redis
channels-redis==2.4.2
    # via -r requirements-server.in
chardet==5.1.0
    # via ods-tools
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
constantly==15.1.0
    # via twisted
coreapi==2.3.3
    # via
    #   -r requirements-server.in
    #   drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==39.0.1
    # via
    #   autobahn
    #   azure-storage-blob
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
    #   service-identity
daphne==2.5.0
    # via channels
django==3.2.18
    # via
    #   channels
    #   django-debug-toolbar
    #   django-filter
    #   django-model-utils
    #   django-request-logging
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-yasg
    #   mozilla-django-oidc
django-celery-results==2.4.0
    # via -r requirements-server.in
django-cleanup==7.0.0
    # via -r requirements-server.in
django-debug-toolbar==3.8.1
    # via -r requirements-server.in
django-filter==22.1
    # via -r requirements-server.in
django-model-utils==4.3.1
    # via -r requirements-server.in
django-request-logging==0.7.5
    # via -r requirements-server.in
django-storages[azure]==1.13.2
    # via -r requirements-server.in
djangorestframework==3.14.0
    # via
    #   -r requirements-server.in
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-yasg
djangorestframework-simplejwt==5.2.2
    # via -r requirements-server.in
drf-nested-routers==0.93.4
    # via -r requirements-server.in
drf-yasg==1.21.5
    # via -r requirements-server.in
greenlet==2.0.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements-server.in
hiredis==2.2.2
    # via aioredis
httplib2==0.21.0
    # via pyrabbit
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via twisted
inflection==0.5.1
    # via drf-yasg
isodate==0.6.1
    # via azure-storage-blob
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.2.0
    # via -r requirements-server.in
josepy==1.13.0
    # via mozilla-django-oidc
jsonpickle==3.0.1
    # via -r requirements-server.in
jsonschema==4.17.3
    # via -r requirements-server.in
kombu==5.2.4
    # via celery
markdown==3.4.1
    # via -r requirements-server.in
markupsafe==2.1.2
    # via jinja2
mozilla-django-oidc==3.0.0
    # via -r requirements-server.in
msgpack==0.6.2
    # via channels-redis
numpy==1.24.2
    # via
    #   pandas
    #   pyarrow
ods-tools==3.0.3
    # via -r requirements-server.in
packaging==23.0
    # via drf-yasg
pandas==1.5.3
    # via
    #   -r requirements-server.in
    #   ods-tools
pathlib2==2.3.7.post1
    # via -r requirements-server.in
prompt-toolkit==3.0.37
    # via click-repl
psycopg2-binary==2.9.5
    # via -r requirements-server.in
pyarrow==11.0.0
    # via -r requirements-server.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pymysql==1.0.2
    # via -r requirements-server.in
pyopenssl==23.0.0
    # via
    #   josepy
    #   twisted
pyparsing==3.0.9
    # via httplib2
pyrabbit==1.1.0
    # via -r requirements-server.in
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   botocore
    #   pandas
pytz==2022.7.1
    # via
    #   celery
    #   django
    #   djangorestframework
    #   drf-yasg
    #   pandas
redis==4.5.1
    # via -r requirements-server.in
requests==2.28.2
    # via
    #   azure-core
    #   coreapi
    #   mozilla-django-oidc
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
s3transfer==0.6.0
    # via boto3
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   azure-core
    #   click-repl
    #   isodate
    #   pathlib2
    #   python-dateutil
    #   service-identity
sqlalchemy==1.4.46
    # via -r requirements-server.in
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
twisted[tls]==22.10.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.5.0
    # via
    #   azure-core
    #   azure-storage-blob
    #   twisted
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.14
    # via
    #   botocore
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
whitenoise==6.4.0
    # via -r requirements-server.in
zope-interface==5.5.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
